contribution guidelines
:heavy_check_mark: added contribution guidelines
:heavy_check_mark: remove some unnecessary files
:heavy_check_mark: cleanup README
