+++ postprocessing_variant_calls/maf/traceback/traceback_process.py
@@ -1,0 +1,60 @@
+#!/usr/bin/env python
+# imports
+import os
+import sys
+from pathlib import Path
+from typing import List, Optional
+import typer
+from postprocessing_variant_calls.maf.helper import (
+    check_maf,
+    check_txt,
+    check_separator,
+    read_tsv,
+    MAFFile,
+    gen_id_tsv,
+)
+import pandas as pd
+import numpy as np
+from typing import Tuple
+
+
+app = typer.Typer(help="post-processing command tagging maf files")
+# app.add_typer(app, name="annotate", help="annotate maf files based on a given input. Currently supports bed and maf files as references.")
+
+@app.command(
+    "common_variant",
+    help="Tag a variant in a MAF file as common variant based on GNOMAD AF",
+)
+def common_variant(
+    maf: Path = typer.Option(
+        ...,
+        "--maf",
+        "-m",
+        exists=True,
+        file_okay=True,
+        dir_okay=False,
+        writable=False,
+        readable=True,
+        resolve_path=True,
+        help="MAF file to subset",
+    ),
+    output_maf: Path = typer.Option(
+        "output.maf", "--output", "-o", help="Maf output file name."
+    ),
+    separator: str = typer.Option(
+        "tsv",
+        "--separator",
+        "-sep",
+        help="Specify a seperator for delimited data.",
+        callback=check_separator,
+    ),
+):
+    # prep maf
+    mafa = MAFFile(maf, separator)
+    mafa = mafa.tag("common_variant")
+    mafa.to_csv(f"{output_maf}".format(outputFile=output_maf), index=False, sep="\t")
+    return 0
+
+
+if __name__ == "__main__":
+    app()
